Solution to Check Triangle Validity

+++ Check Triangle Validity@@6788dd63490637cfc44e1b93/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    int h,p,b;
-    scanf("%d %d %d",&h,&b,&p);
+    float h,p,b;
+    scanf("%f %f %f",&h,&b,&p);
     if(b*b+p*p==h*h){
         printf("Valid");
     }
